Update Database section in settings.py
This update allows for better parsing of (multiple) lines defined in
`.pgpass`. The script now reads the database name and compares it to a
script-defined variable, rather than simply taking the password from a
hard-coded offset. Now, all credentials (host, port, database name,
username, and password) are read from each line. If the credentials for
the database are not found, a custom exception, CredentialsNotFound, is
raised.