Reroute navbar title link and rename Home link
It was getting annoying clicking on the title and being directed to the
homepage. Since the homepage is nothing more than a visual landing page
that only needs to appear when the user first connects to the site, I
imagine users would expect to be taken to the Gallery page after
clicking the title. With that implemented, having a link that says
"Home" point to a different page than the title is confusing, so it's
been renamed to "Title Page" instead.