Fix infinite scroll bug w/ page updating too late
There was an issue where scrolling down the page would allow the bottom
to be seen, but after hitting the bottom and scrolling up a bit would
cause the next portion to be loaded. The page was usable, but scrolling
past a certain point proved inconvenient to the user. This commit aims
to correct that behavior by loading the page when the top of the last
loaded element becomes visible, and do so in a way that is functionally
independent from the scroll-to-top button.

Rather than checking if the entire element is visible, just check if it
is partially visible. To achieve this, simply remove the first half of
the `&&` statement in `isInViewport`. Also remove the `elementBottom`
variable, and change the remaining variables to constants as they never
need to be redefined. Now the GET request for the next page in the
infinite scroll series can be triggered earlier, reducing the
likelihood of hitting the bottom of the page.

Use separate booleans to keep track of infinite scroll GET request and
scroll-to-top animation events. Before, this was tracked with a single
`cooldown` variable that was updated with a timeout, but in both cases,
it is much better to simply update those newly created variables at the
end of a callback.